split crafting panel
